Added HEAD and OPTIONS support to DefaultHandler

+++ src/jaws/business/defaultmodule/DefaultHandler.java
@@ -10,6 +10,7 @@ import java.util.Map;
 
 import jaws.module.http.HTTPRequest;
 import jaws.module.http.HTTPResponse;
+import jaws.module.http.RequestMethod;
 import jaws.module.net.Handle;
 
 public class DefaultHandler {
@@ -41,35 +42,45 @@ public class DefaultHandler {
 	@Handle(extensions = {".*"}, priority = Integer.MIN_VALUE)
 	public static HTTPResponse handle(HTTPRequest request, HTTPResponse response, File webRoot) throws IOException {
 
-		try {
-			File file = new File(webRoot, request.url().substring(1));
-			// if the requested path is a folder, try to get the 'index.html' file
-			if(file.isDirectory() && !new File(file, "index.html").exists()) {
-				String body = "";
-				String[] fileNames = file.list();
-				for(String fileName : fileNames) {
-					body += "<a href=\"" + request.url() + "/" + fileName + "\">" + fileName + (new File(file, fileName).isDirectory()?"/":"") + "</a><br>";
-				}
-
-				response.body(body);
+		File file = new File(webRoot, request.url().substring(1));
+		if (file.exists()) {
+			if (request.method() == RequestMethod.OPTIONS) {
+				
+				response.statusCode(200)
+				        .header("Allow", "GET, HEAD, OPTIONS");
 			} else {
-				if(file.isDirectory() && new File(file, "index.html").exists()) {
-					file = new File(file, "index.html");
-				}
-				//String content = new String(Files.readAllBytes(file.toPath()));
-				ByteArrayOutputStream out = new ByteArrayOutputStream();
-				response.body(out);
-				Files.copy(file.toPath(), out);
-				String fileExtension = request.url().substring(request.url().lastIndexOf('.') + 1);
-				if(mimeTypes.containsKey(fileExtension)) {
-					response.header("Content-Type", mimeTypes.get(fileExtension));
+				
+				// if the requested path is a folder, try to get the 'index.html' file
+				if(file.isDirectory() && !new File(file, "index.html").exists() && request.method() != RequestMethod.HEAD) {
+					
+					String body = "";
+					String[] fileNames = file.list();
+					for(String fileName : fileNames) {
+						body += "<a href=\"" + request.url() + "/" + fileName + "\">" + fileName + (new File(file, fileName).isDirectory()?"/":"") + "</a><br>";
+					}
+	
+					response.body(body);
+				} else {
+					
+					if(file.isDirectory() && new File(file, "index.html").exists()) {
+						file = new File(file, "index.html");
+					}
+					
+					if (request.method() != RequestMethod.HEAD) {
+						
+						ByteArrayOutputStream out = new ByteArrayOutputStream();
+						response.body(out);
+						Files.copy(file.toPath(), out);
+					}
+					String fileExtension = request.url().substring(request.url().lastIndexOf('.') + 1);
+					if(mimeTypes.containsKey(fileExtension)) {
+						response.header("Content-Type", mimeTypes.get(fileExtension));
+					}
 				}
 			}
-		}  catch(NoSuchFileException e) {
+		} else {
 			response.statusCode(404).reason("Not Found").body("<h1>404 - Not Found</h1>");
 		}
-
-		//System.out.println(response.body());
 
 		return response.header("Content-Length", Integer.toString(response.body().length));
 	}
